requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
abseil-cpp=20230802.0=h6a678d5_2
absl-py=1.4.0=py312h06a4308_0
aom=3.6.0=h6a678d5_0
appdirs=1.4.4=pyhd3eb1b0_0
asttokens=2.0.5=pyhd3eb1b0_0
blas=1.0=mkl
blosc=1.21.3=h6a678d5_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brunsli=0.1=h2531618_0
bzip2=1.0.8=h5eee18b_5
c-ares=1.19.1=h5eee18b_0
ca-certificates=2024.3.11=h06a4308_0
certifi=2024.2.2=py312h06a4308_0
cfitsio=3.470=h5893167_7
charls=2.2.0=h2531618_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=py312h06a4308_0
colorama=0.4.6=py312h06a4308_0
comm=0.2.1=py312h06a4308_0
contourpy=1.2.0=py312hdb19cb5_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cycler=0.11.0=pyhd3eb1b0_0
cyrus-sasl=2.1.28=h52b45da_1
dav1d=1.2.1=h5eee18b_0
dbus=1.13.18=hb2f20db_0
debugpy=1.6.7=py312h6a678d5_0
decorator=5.1.1=pyhd3eb1b0_0
docker-pycreds=0.4.0=pyhd3eb1b0_0
efficientnet-pytorch=0.7.1=pyhd8ed1ab_1
executing=0.8.3=pyhd3eb1b0_0
expat=2.6.2=h59595ed_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py312h06a4308_0
fontconfig=2.14.2=h14ed4e7_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
fsspec=2023.10.0=py312h06a4308_0
giflib=5.2.1=h5eee18b_3
gitdb=4.0.7=pyhd3eb1b0_0
gitpython=3.1.37=py312h06a4308_0
glib=2.78.4=h6a678d5_0
glib-tools=2.78.4=h6a678d5_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
grpc-cpp=1.48.2=he1ff14a_4
grpcio=1.48.2=py312he1ff14a_4
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
gtest=1.14.0=hdb19cb5_0
huggingface_hub=0.20.3=py312h06a4308_0
icu=73.1=h6a678d5_0
idna=3.4=py312h06a4308_0
imagecodecs=2023.1.23=py312h81b8100_1
imageio=2.33.1=py312h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.29.3=pyhd33586a_0
ipython=8.20.0=py312h06a4308_0
jedi=0.18.1=py312h06a4308_1
jinja2=3.1.3=py312h06a4308_0
jpeg=9e=h5eee18b_1
jupyter_client=8.6.0=py312h06a4308_0
jupyter_core=5.5.0=py312h06a4308_0
jxrlib=1.1=h7b6447c_2
kiwisolver=1.4.4=py312h6a678d5_0
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lazy_loader=0.3=py312h06a4308_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libaec=1.0.4=he6710b0_1
libavif=0.11.1=h5eee18b_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.9.1.3=0
libcups=2.4.2=h2d74bed_1
libcurand=10.3.5.147=0
libcurl=8.5.0=h251f7ec_0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libexpat=2.6.2=h59595ed_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=13.2.0=h807b86a_5
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libglib=2.78.4=hdc74915_0
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libllvm14=14.0.6=hdb19cb5_3
libnghttp2=1.57.0=h2d74bed_0
libnpp=11.8.0.86=0
libnsl=2.0.1=hd590300_0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libpq=12.17=hdbd6064_0
libprotobuf=3.20.3=he621ea3_0
libsodium=1.0.18=h7b6447c_0
libsqlite=3.45.3=h2797004_0
libssh2=1.10.0=hdbd6064_2
libstdcxx-ng=13.2.0=h7e041cc_5
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=2.38.1=h0b41bf4_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxcrypt=4.4.36=hd590300_1
libxkbcommon=1.0.1=h5eee18b_1
libxml2=2.10.4=hf1b16e4_1
libzlib=1.2.13=hd590300_5
libzopfli=1.0.3=he6710b0_0
lightning-utilities=0.9.0=py312h06a4308_0
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.1=py312h06a4308_0
markupsafe=2.1.3=py312h5eee18b_0
matplotlib=3.8.4=py312h06a4308_0
matplotlib-base=3.8.4=py312h526ad5a_0
matplotlib-inline=0.1.6=py312h06a4308_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py312h5eee18b_1
mkl_fft=1.3.8=py312h5eee18b_0
mkl_random=1.2.4=py312hdb19cb5_0
mpmath=1.3.0=py312h06a4308_0
munch=2.5.0=pyhd3eb1b0_0
munkres=1.1.4=py_0
mysql=5.7.24=h721c034_2
ncurses=6.4.20240210=h59595ed_0
nest-asyncio=1.6.0=py312h06a4308_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py312h06a4308_0
numpy=1.26.4=py312hc5e2394_0
numpy-base=1.26.4=py312h0da6c21_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.2.1=hd590300_1
packaging=23.2=py312h06a4308_0
parso=0.8.3=pyhd3eb1b0_0
pathtools=0.1.2=pyhd3eb1b0_1
pcre2=10.42=hebb0a14_0
pexpect=4.8.0=pyhd3eb1b0_3
pillow=10.2.0=py312h5eee18b_0
pip=23.3.1=py312h06a4308_0
platformdirs=3.10.0=py312h06a4308_0
ply=3.11=py312h06a4308_1
pretrainedmodels=0.7.4=pyhd8ed1ab_2
prompt-toolkit=3.0.43=py312h06a4308_0
prompt_toolkit=3.0.43=hd3eb1b0_0
protobuf=3.20.3=py312h6a678d5_0
psutil=5.9.0=py312h5eee18b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pydensecrf=1.0rc3=py312hfb8ada1_5
pygments=2.15.1=py312h06a4308_1
pyparsing=3.0.9=py312h06a4308_0
pyqt=5.15.10=py312h6a678d5_0
pyqt5-sip=12.13.0=py312h5eee18b_0
python=3.12.3=hab00c5b_0_cpython
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.12=4_cp312
pytorch=2.2.2=py3.12_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-lightning=2.2.2=pyhd8ed1ab_0
pytorch-mutex=1.0=cuda
pyyaml=6.0.1=py312h5eee18b_0
pyzmq=25.1.2=py312h6a678d5_0
qt-main=5.15.2=h53bd1ea_10
re2=2022.04.01=h295c915_0
readline=8.2=h5eee18b_0
requests=2.31.0=py312h06a4308_1
safetensors=0.4.3=py312h4b3b743_0
scikit-image=0.22.0=py312h526ad5a_0
scipy=1.12.0=py312hc5e2394_0
segmentation-models-pytorch=0.3.3=pyhd8ed1ab_0
sentry-sdk=1.9.0=py312h06a4308_0
setproctitle=1.2.2=py312h5eee18b_0
setuptools=68.2.2=py312h06a4308_0
sip=6.7.12=py312h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
smmap=4.0.0=pyhd3eb1b0_0
snappy=1.1.10=h6a678d5_1
sqlite=3.41.2=h5eee18b_0
stack_data=0.2.0=pyhd3eb1b0_0
sympy=1.12=py312h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.16.2=pyhd8ed1ab_0
tensorboard-data-server=0.7.0=py312h52d8a92_0
tensorboardx=2.6.2.2=pyhd8ed1ab_0
tifffile=2023.4.12=py312h06a4308_0
timm=0.9.2=pyhd8ed1ab_0
tk=8.6.13=noxft_h4845f30_101
torchaudio=2.2.2=py312_cu118
torchmetrics=1.3.2=pyhd8ed1ab_0
torchvision=0.17.2=py312_cu118
tornado=6.3.3=py312h5eee18b_0
tqdm=4.66.2=pyhd8ed1ab_0
traitlets=5.7.1=py312h06a4308_0
typing-extensions=4.9.0=py312h06a4308_1
typing_extensions=4.9.0=py312h06a4308_1
tzdata=2024a=h04d1e81_0
unzip=6.0=h611a1e1_0
urllib3=2.1.0=py312h06a4308_0
wandb=0.16.5=pyhd8ed1ab_0
wcwidth=0.2.5=pyhd3eb1b0_0
werkzeug=2.3.8=py312h06a4308_0
wheel=0.41.2=py312h06a4308_0
xz=5.4.6=h5eee18b_0
yaml=0.2.5=h7b6447c_0
zeromq=4.3.5=h6a678d5_0
zfp=1.0.0=h6a678d5_0
zlib=1.2.13=hd590300_5
zstd=1.5.5=hc292b87_0
